misc/web-abat-xwalk-tests/abat/common/deviceorientation/DeviceOrientationEvent_attributes_test.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Xinx, liu <xinx.liu@intel.com>

-->

<meta charset='utf-8'>
<title>Orientation Test: DeviceOrientationEvent attributes exists and type are right</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/TR/2011/WD-orientation-event-20111201/#deviceorientation">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  setup(function(){
    tizen.power.turnScreenOn();
  }, {explicit_done: true, timeout: 2000});
  var run = false;
  on_event(window, "deviceorientation", function(event) {
    if(!run) {
      run = true;
      test(function() {
        assert_true("absolute" in event, "Attribute absolute missing in DeviceOrientationEvent object");
      }, "Check if the DeviceOrientationEvent.absolute attribute exists");

      test(function() {
        assert_equals(typeof event.absolute, "boolean", "Type of attribute absolute");
      }, "Check if DeviceOrientationEvent.absolute is type of boolean");

      test(function() {
        assert_true("alpha" in event, "Attribute alpha missing in DeviceOrientationEvent object");
      }, "Check if the DeviceOrientationEvent.alpha attribute exists");

      test(function() {
        assert_equals(typeof event.alpha, "number", "Attribute alpha is type of");
      }, "Check if DeviceOrientationEvent.alpha is type of double");

      test(function() {
        assert_true("beta" in event, "Attribute beta missing in DeviceOrientationEvent object");
      }, "Check if the DeviceOrientationEvent.beta attribute exists");

      test(function() {
        assert_equals(typeof event.beta, "number", "Attribute beta is type of");
      }, "Check if DeviceOrientationEvent.beta is type of double");

      test(function() {
        assert_true("gamma" in event, "Attribute gamma missing in DeviceOrientationEvent object");
      }, "Check if the DeviceOrientationEvent.gamma attribute exists");

      test(function() {
        assert_equals(typeof event.gamma, "number", "Attribute gamma is type of");
      }, "Check if DeviceOrientationEvent.gamma is type of double");
      done();
    }
  });
  setTimeout(function () {
    test(function() {
      assert_unreached("Can not fire deviceorientation event");
    }, "Check if the DeviceOrientationEvent.absolute attribute exists");
    test(function() {
      assert_unreached("Can not fire deviceorientation event");
    }, "Check if DeviceOrientationEvent.absolute is type of boolean");
    test(function() {
      assert_unreached("Can not fire deviceorientation event");
    }, "Check if the DeviceOrientationEvent.alpha attribute exists");
    test(function() {
      assert_unreached("Can not fire deviceorientation event");
    }, "Check if DeviceOrientationEvent.alpha is type of double");
    test(function() {
      assert_unreached("Can not fire deviceorientation event");
    }, "Check if the DeviceOrientationEvent.beta attribute exists");
    test(function() {
      assert_unreached("Can not fire deviceorientation event");
    }, "Check if DeviceOrientationEvent.beta is type of double");
    test(function() {
      assert_unreached("Can not fire deviceorientation event");
    }, "Check if the DeviceOrientationEvent.gamma attribute exists");
    test(function() {
      assert_unreached("Can not fire deviceorientation event");
    }, "Check if DeviceOrientationEvent.gamma is type of double");
    done();
  }, 1500)
</script>
